chore: refactoring run pt1, types, typos and dedup
